Move coalesce
Ref #2698
