+++ core/base-service/coalesce.spec.js
