+++ lib/logos.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { toSvgColor } = require('../gh-badges/lib/color')
-const coalesce = require('./coalesce')
+const coalesce = require('../core/base-service/coalesce')
 const { svg2base64 } = require('./svg-helpers')
 
 const logos = require('./load-logos')()
